refactor: clean up survey123 controller/util
