+++ src/utils/index.js
@@ -1,5 +1,4 @@
 import {
-  cleanCsvCreator,
   csvDownloadCreator,
   deleteFile,
   processCSV,
@@ -33,7 +32,6 @@ import {
 } from './pipeline';
 
 import {
-  csvUploadSurvey123Creator,
   deleteInsert,
   survey123UnpackCreator,
   survey123WebhookUnpackCreator,
@@ -42,9 +40,7 @@ import {
 
 export {
   calculatedFieldsGeneratorCreator,
-  cleanCsvCreator,
   csvDownloadCreator,
-  csvUploadSurvey123Creator,
   deleteFile,
   deleteInsert,
   extractObjectFieldsCreator,
